Move modals to dedicated component
