+++ src/App.svelte
@@ -9,25 +9,15 @@
 	import { 
 		animationType, footerData, osSelectOptions, 
 		installConfigData, manualDownloadGroup, manualDownloadFileData, 
-		eventObject, showSelectDeviceModal, showDeveloperModeModal,
-		deviceSelectOptions, userActionEventObject, actionData,
+		eventObject,
+		userActionEventObject, actionData,
 		deviceName, osInstructsData
 	} from './stores.mjs';
 
 	//Modals
-	import NewUpdateModal from './ui/modals/NewUpdateModal.svelte'
-	import UdevModal from './ui/modals/UdevModal.svelte'
-	import ConnectionLostModal from './ui/modals/ConnectionLostModal.svelte'
-	import NoNetworkModal from './ui/modals/NoNetworkModal.svelte'
-	import LowPowerModal from './ui/modals/LowPowerModal.svelte'
-	import WindowsDriversModal from './ui/modals/WindowsDriversModal.svelte'
-	import SelectDeviceModal from './ui/modals/SelectDeviceModal.svelte'
-	import DeveloperModeModal from './ui/modals/DeveloperModeModal.svelte'
-	import ErrorModal from './ui/modals/ErrorModal.svelte'
-	import ResultModal from './ui/modals/ResultModal.svelte'
-	import UnlockModal from './ui/modals/UnlockModal.svelte'
-	import OptionsModal from './ui/modals/OptionsModal.svelte'
-	import OemLockModal from './ui/modals/OemLockModal.svelte'
+	import Modals from './ui/modals/Modals.svelte'
+	import OptionsModal from './ui/modals/specific-modals/OptionsModal.svelte'
+	import ResultModal from './ui/modals/specific-modals/ResultModal.svelte'
 
 	//Routing
 	import Router from 'svelte-spa-router'
@@ -40,26 +30,9 @@
 	global.packageInfo = remote.getGlobal("packageInfo");
 
 	//Modal variables
-	let showNewUpdateModal = false;
-	let showUdevModal = false;
-	let showConnectionLostModal = false;
-	let showNoNetworkModal = false;
-	let showLowPowerModal = false;
-	let showWindowsDriversModal = false;
-	let show_selectDeviceModal;
-	let show_developerModeModal = false;
-	let showErrorModal = false;
-	let showResultModal = false;
-	let showUnlockModal = false;
 	let showOptionsModal = false;
-	let showOemLockModal = false;
-	
-	//Modal props
-	let select_options;
-	let errorData;
+	let showResultModal = false;
 	let showDoNotAskAgainButton;
-	let unlockData;
-	let oemUnlockData;
 
 	//Footer
 	let progressBarWidth = 0;
@@ -68,41 +41,22 @@
 
 	ipcRenderer.on("user:write:status", (e, status, waitDots) => {
 		//footer.topText.set(status, waitDots);
-		// footerData.set({
-		// 	topText: status,
-		// });
-		// footerData.topText.set({
-		// 	status
-		// });
+		$footerData.topText = status;
 	});
 
 	ipcRenderer.on("user:write:under", (e, status) => {
 		//footer.underText.set(status, true);
-		// footerData.underText.set({
-		// 	status
-		// });
+		$footerData.underText = status;
 	});
 
 	ipcRenderer.on("user:write:speed", (e, speed) => {
-		//footer.speedText.set(speed);
+		$footerData.speedText = speed;
 	});
 
 	//Reactive variables
 	const unsubscribeFooterData = footerData.subscribe(value => {
     footer_data = value;
 	});
-	
-	const unsubscribeShowSelectDeviceModal = showSelectDeviceModal.subscribe(value => {
-    show_selectDeviceModal = value;
-	});
-	
-	const unsubscribeShowDeveloperModeModal = showDeveloperModeModal.subscribe(value => {
-    show_developerModeModal = value;
-	});
-	
-	const unsubscribeDeviceSelectOptions = deviceSelectOptions.subscribe(value => {
-    select_options = value;
-  });
 
 	//Life cycle methods
 	onMount(() => {
@@ -113,48 +67,6 @@
 	});
 
 	//Messages
-	//Modal related messages
-	if (process.platform === "linux" && !process.env.SNAP) {
-		ipcRenderer.invoke("getSettingsValue", "never.udev")
-			.then(never => {
-				if (never) {
-					setTimeout(() => showUdevModal = true, 1000)
-			}
-		})
-	}
-
-	if (process.platform === "win32") {
-		ipcRenderer.invoke("getSettingsValue", "never.windowsDrivers")
-			.then(never => {
-				if (never) {
-					setTimeout(() => showWindowsDriversModal = true, 1000)
-				}
-			})
-	}
-
-	ipcRenderer.on("user:connection-lost", () => {
-		showConnectionLostModal = true;
-	});
-	
-	ipcRenderer.on("user:no-network", () => {
-    showNoNetworkModal = true;
-	});
-	
-	ipcRenderer.on("user:low-power", (callback) => {
-    showLowPowerModal = true;
-	});
-
-	ipcRenderer.on("user:update-available", () => {
-      showNewUpdateModal = true;
-	});
-
-	ipcRenderer.on("user:error", (event, error, restart, ignore) => {
-		errorData = error;
-		showErrorModal = true;
-	});
-
-	ipcRenderer.on("user:report", (_, done) => requestReport(done));
-
 	//Routing messages
 	ipcRenderer.on("user:write:working", (e, animation) => {
 		animationType.set(animation);
@@ -198,8 +110,7 @@
 		});
 
 		osSelectOptions.set(osSelects);
-		installConfigData.set(installConfig)
-		
+		installConfigData.set(installConfig)	
 		push('/select-os');
 	});
 
@@ -220,41 +131,20 @@
 		});
 		showOptionsModal = true;
 	});
-	
-	ipcRenderer.on("user:eula", (_, eula) => {
-		unlockData = eula;
-		showUnlockModal = true;
-		console.log("eula", eula);
-	});
 
-	ipcRenderer.on("user:unlock", (_, unlock, user_actions) => {
-		unlockData = user_actions
-		showUnlockModal = true;
-	});
+	ipcRenderer.on("user:report", (_, done) => requestReport(done));
 
-	ipcRenderer.on("user:prerequisites", (_, prerequisites, user_actions) => {
-		unlockData = user_actions
-		showUnlockModal = true;
-	});
-	
-	ipcRenderer.on("user:oem-lock", (event, enable = false, url) => {
-		oemUnlockData = {
-			enable: enable,
-			url: url,
-		};
-		showOemLockModal = true;
-  });
+
+	//Other methods
+	function requestReport(done = false) {
+		done? showDoNotAskAgainButton = true : showDoNotAskAgainButton = true;
+		showResultModal = true;
+	}
 
 	//Error handling
 	// Catch all unhandled errors in rendering process
 	window.onerror = (err, url, line) => {
 		ipcRenderer.send("renderer:error", err + " (MainRenderer:" + line + ")");
-	}
-		
-	//Other methods
-	function requestReport(done = false) {
-		done? showDoNotAskAgainButton = true : showDoNotAskAgainButton = true;
-		showResultModal = true;
 	}
 </script>
 
@@ -270,44 +160,12 @@
 	</div>
 	<div class="view-container container">
 		<Router {routes}/>
-		{#if showNewUpdateModal}
-		<NewUpdateModal on:close={() => showNewUpdateModal = false}/>
-		{/if}
-		{#if showUdevModal}
-		<UdevModal on:close={() => showUdevModal = false}/>
-		{/if}
-		{#if showConnectionLostModal}
-		<ConnectionLostModal on:close={() => showConnectionLostModal = false}/>
-		{/if}
-		{#if showNoNetworkModal}
-		<NoNetworkModal on:close={() => showNoNetworkModal = false}/>
-		{/if}
-		{#if showLowPowerModal}
-		<LowPowerModal on:close={() => showLowPowerModal = false}/>
-		{/if}
-		{#if showWindowsDriversModal}
-		<WindowsDriversModal on:close={() => showWindowsDriversModal = false}/>
-		{/if}
-		{#if show_selectDeviceModal}
-		<SelectDeviceModal selectOptions={select_options} on:close={() => showSelectDeviceModal.set(false)}/>
-		{/if}
-		{#if show_developerModeModal}
-		<DeveloperModeModal on:close={() => showDeveloperModeModal.set(false)}/>
-		{/if}
-		{#if showErrorModal}
-		<ErrorModal errorData={errorData} on:close={() => showErrorModal = false}/>
-		{/if}
-		{#if showResultModal}
-		<ResultModal showDoNotAskAgainButton={showDoNotAskAgainButton} on:close={() => showResultModal = false}/>
-		{/if}
-		{#if showUnlockModal}
-		<UnlockModal unlockData={unlockData} on:close={() => showUnlockModal = false}/>
-		{/if}
+		<Modals/>
 		{#if showOptionsModal}
 		<OptionsModal on:close={() => showOptionsModal = false}/>
 		{/if}
-		{#if showOemLockModal}
-		<OemLockModal oemUnlockData={oemUnlockData} on:close={() => showOemLockModal = false}/>
+		{#if showResultModal}
+		<ResultModal showDoNotAskAgainButton={showDoNotAskAgainButton} on:close={() => showResultModal = false}/>
 		{/if}
 	</div>
 	<div class="progress">
@@ -344,6 +202,7 @@
 		flex-direction: row;
 		justify-content: space-between;
 		align-items: center;
+		height: 60px;
 		max-height: 60px;
 		padding: 0 10px 0 10px;
 		background-color: #f5f5f5;
@@ -357,6 +216,7 @@
 		display: flex;
 		flex: 1 1 auto;
 		flex-direction: column;
+		padding: 20px 10px;
 	}
 
 	.progress {
@@ -378,6 +238,7 @@
 		display: flex;
 		flex: 1 1 auto;
 		flex-direction: row;
+		height: 90px;
 		max-height: 90px;
 		background-color: #f5f5f5;
 		margin: 0;
