+++ src/css/style.css
@@ -1,60 +1,3 @@
-html {
-  position: relative;
-  min-height: 100%;
-}
-body {
-  margin-bottom: 60px;
-  overflow:hidden;
-  -webkit-font-smoothing: antialiased;
-  text-rendering: optimizeLegibility;
-  image-rendering: crisp-edges;
-}
-body .vertical-centered-box {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  z-index: 6;
-}
-body .vertical-centered-box:after {
-  content: '';
-  display: inline-block;
-  height: 100%;
-  vertical-align: middle;
-  margin-right: -0.25em;
-  z-index: 6;
-}
-body .vertical-centered-box .content {
-  -webkit-box-sizing: border-box;
-  -moz-box-sizing: border-box;
-  box-sizing: border-box;
-  display: inline-block;
-  vertical-align: middle;
-  text-align: left;
-  font-size: 0;
-  z-index: 6;
-}
-
-.help-button {
-  z-index: 10 !important;
-  top: 15px !important;
-  position: absolute;
-  right: 15px !important;
-  height: 30px !important;
-  margin-bottom: 0px !important;
-  margin-right: 0px !important;
-}
-
-.donate-button {
-  z-index: 10 !important;
-  top: 15px !important;
-  position: absolute;
-  right: 135px !important;
-  height: 30px !important;
-  margin-bottom: 0px !important;
-  margin-right: 0px !important;
-}
-
 .ubp-robot-foot {
   bottom: 10px !important;
   position: absolute;
@@ -102,55 +45,6 @@ body .vertical-centered-box .content {
   z-index: 6;
 }
 
-.is-over {
-  z-index: 6;
-}
-
-.is-under {
-  z-index: -2;
-}
-
-.loader-circle {
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  width: 200px;
-  height: 200px;
-  border-radius: 50%;
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.1);
-  background-color: white;
-  margin-left: -100px;
-  margin-top: -100px;
-  z-index: 6;
-}
-.loader-line-mask {
-  position: absolute;
-  left: 50%;
-  top: 50%;
-  width: 100px;
-  height: 200px;
-  margin-left: -100px;
-  margin-top: -100px;
-  overflow: hidden;
-  -webkit-transform-origin: 100px 100px;
-  -moz-transform-origin: 100px 100px;
-  -ms-transform-origin: 100px 100px;
-  -o-transform-origin: 100px 100px;
-  transform-origin: 100px 100px;
-  -webkit-mask-image: -webkit-linear-gradient(top, #000000, rgba(0, 0, 0, 0));
-  -webkit-animation: rotate 1.2s infinite linear;
-  -moz-animation: rotate 1.2s infinite linear;
-  -o-animation: rotate 1.2s infinite linear;
-  animation: rotate 1.2s infinite linear;
-  z-index: 6;
-}
-.loader-line-mask .loader-line {
-  width: 200px;
-  height: 200px;
-  border-radius: 50%;
-  box-shadow: inset 0 0 0 1px rgba(0, 0, 0, 0.5);
-  z-index: 6;
-}
 #particles-background,
 #particles-foreground {
   left: -51%;
@@ -253,42 +147,7 @@ lesshat-selector {
 [not-existing] {
   zoom: 1;
 }
-.progress {
-  position: absolute;
-  bottom: 90px;
-  width: 0%;
-  height: 4px;
-  background-color: #E95420;
-  margin: 0;
-  z-index: 6
-}
-.footer {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  height: 90px;
-  background-color: #f5f5f5;
-  margin: 0;
-}
-.container {
-  width: auto;
-  max-width: 680px;
-  padding: 20px 10px;
-}
-.text-muted .installer {
-  display: inline-block;
-}
-.footer-top {
-  margin-top: 0px;
-}
-.header {
-  display: flex;
-  height: 60px;
-  background-color: #f5f5f5;
-  position: fixed;
-  width: 100%;
-  top: 0;
-}
+
 .header-image {
   width: 120px;
   height: auto;
@@ -314,15 +173,4 @@ lesshat-selector {
   position: absolute;
   bottom: 0;
   left: 0;
-}
-
-.modal .close {
-  display: none;
-}
-
-@font-face {
-  font-family: 'Ubuntu';
-  font-style: normal;
-  font-weight: 400;
-  src: local('Ubuntu'), local('Ubuntu Regular'), local('Ubuntu-Regular'), url('../fonts/Ubuntu.ttf') format('truetype');
 }
