--- src/pug/index.pug
@@ -1,65 +1,0 @@
-doctype html
-html
-  head
-    style
-      include ../../node_modules/bootstrap/dist/css/bootstrap.min.css
-    style
-      include ../css/style.css
-  body
-    script
-      include scripts/root.js
-    .header.is-over
-      h3#header-text.text-muted.installer UBports Installer (#{installerVersion})
-    button#help.help-button.btn.btn-primary(onclick="requestReport()") Report a bug
-    button#donate.donate-button.btn.btn-primary Donate
-    #particles-background.vertical-centered-box
-    #particles-foreground.vertical-centered-box
-    #download-animation(style='display: none').vertical-centered-box.download-animation.is-under
-      i.is-under
-    #push-animation(style='display: none').vertical-centered-box.push-animation.is-under
-      i.is-under
-    // Views
-    include views/done
-    include views/user-action
-    include views/manual-download
-    include views/select-os
-    include views/not-supported
-    include views/wait-for-device
-    include views/working
-    #progress.progress(hidden='hidden')
-    footer.footer.is-over
-      .container
-        h3(style="margin-top: 0").text-muted.footer-top
-          span#footer-top UBports Installer is starting up
-          span#wait-dot
-        p
-          span#footer-bottom.text-muted Starting adb service
-          span#footer-speed.text-muted
-    // --- Modals ---
-    // low prio
-    include modals/select-device
-    include modals/oem-lock
-    include modals/developer-mode
-    include modals/options
-    // medium prio
-    include modals/unlock
-    include modals/result
-    // high prio
-    include modals/udev
-    include modals/windows-drivers
-    include modals/new-update
-    // errors
-    include modals/low-power
-    include modals/connection-lost
-    include modals/no-network
-    include modals/error
-  // --- Scripts ---
-  // Animated loading screen
-  script
-    include scripts/particleground.js
-  // UI-related event management
-  script
-    include scripts/ui.js
-  // Functionality-related event management
-  script
-    include scripts/dots.js
