--- src/pug/modals/error.pug
@@ -1,32 +1,0 @@
-#error-modal.modal.fade(tabindex='-1', role='dialog')
-  .modal-dialog(role='document')
-    .modal-content
-      .modal-header
-        button.close(type='button', data-dismiss='modal', aria-label='Close')
-          span(aria-hidden='true') ×
-        h4.modal-title Yikes!
-      .modal-body
-        p The Installation stopped due to a problem. You can choose to ignore this, or re-start the installation process.
-        p If this continues to happen, please check if you are affected by  #[a(onclick="shell.openExternal('https://github.com/ubports/ubports-installer/issues')") a known bug].
-        p If your problem is not yet known, click the button below to report a new bug.
-        pre#error-body.
-          \n
-        p#not-latest-stable(hidden='hidden') You are not using the latest stable version of the UBports Installer.
-        #generic-update-instructions-error.p(hidden='hidden') You can #[a(onclick="shell.openExternal('https://github.com/ubports/ubports-installer/releases/latest')") download the latest version from GitHub].
-        #snap-update-instructions-error.p(hidden='hidden') Run #[code snap refresh ubports-installer --stable] in your terminal to install the latest version.
-        p If you need help, you can join UBports' support channels on #[a(onclick="shell.openExternal('https://t.me/WelcomePlus')") telegram] or #[a(onclick="shell.openExternal('https://matrix.to/#/!KwdniMNeTmClpgHkND:matrix.org?via=matrix.org&via=ubports.chat&via=disroot.org')") matrix] or ask a question #[a(onclick="shell.openExternal('https://forums.ubports.com')") in the forum] or on #[a(onclick="shell.openExternal('https://askubuntu.com')") askubuntu]. As a last resort, we also have #[a(onclick="shell.openExternal('https://devices.ubuntu-touch.io')") manual installation instructions for every device], that you can follow if you want to install without using the UBports Installer.
-
-      .modal-footer
-        button.btn.btn-default(type='button', hidden='hidden', data-dismiss='modal', onclick='ipcRenderer.send("user:error:reply", "restart")') Try again
-        button.btn.btn-default(type='button', hidden='hidden', data-dismiss='modal', onclick='ipcRenderer.send("user:error:reply", "ignore")') Ignore
-        button#btn-bugreport.btn.btn-primary(type='button', onclick='ipcRenderer.send("user:error:reply", "bugreport")') Report a bug
-  script.
-    ipcRenderer.on("user:error", (event, error, restart, ignore) => {
-      $("#error-body").text(error);
-      $("#error-modal").modal("show");
-    });
-
-    // Catch all unhandled errors in rendering process
-    window.onerror = (err, url, line) => {
-      ipcRenderer.send("renderer:error", err + " (MainRenderer:" + line + ")");
-    }
