--- src/pug/views/done.pug
@@ -1,22 +1,0 @@
-#views-done.main.container.views(hidden='hidden')
-  .row
-    .col-xs-6
-      img(style='height: 350px; margin: auto; display: block;', src='../screens/Screen6.jpg')
-    .col-xs-6
-      h4(style='font-weight: bold;') Success!
-      p
-        | The installation process from the computer is done. The device will now perform the remaining steps which should take less than five minutes. After the installation, the device will reboot and you can begin to explore Ubuntu Touch. 
-      p
-        | Found something you don't like?  #[a(onclick="shell.openExternal('https://github.com/ubports/ubports-touch')") Tell us], or help us change it!
-      p
-        | Development of Ubuntu Touch is driven by the  #[a(onclick="shell.openExternal('https://ubports.com')") UBports Community]. Donate now to allow us to continue our mission!
-      button.btn.btn-default(type='button', style='width: 49%; margin-bottom: 10px; margin-right: 5px', onclick="shell.openExternal('https://ubports.com/join-us')") Get involved 
-      button.btn.btn-primary(type='button', style='width: 49%; margin-bottom: 10px;', onclick="shell.openExternal('https://ubports.com/donate')") Donate
-      p
-        | Got more devices you want to flash?
-      button.btn.btn-info(type='button', style='width: 100%;', onclick="ipcRenderer.send('restart');") Flash another device!
-  script.
-    ipcRenderer.on("user:write:done", () => {
-      views.show("done");
-      $("#progress").width("0%");
-    });
