--- src/pug/views/manual-download.pug
@@ -1,42 +1,0 @@
-#views-manual-download.main.container.views(hidden='hidden')
-  .row
-    .col-xs-6
-      img(style='height: 350px; margin: auto; display: block;', src='../screens/Screen6.jpg')
-    .col-xs-6
-      h4(style='font-weight: bold;') Manual Download
-      p The #[b#manual-download-group] file #[b#manual-download-filename-1] needs to be manually downloaded due to licensing restrictions. Sorry for the inconvenience!
-      p Please download the file #[b#manual-download-filename-2] from #[a#manual-download-link].
-      p#manual-download-after-instructions Once you have it, select it in the file picker and click the button below to continue.
-      .input-group(style="margin-bottom: 1em;")
-        .custom-file
-          input#manual-download-input.custom-file-input(type="file")
-      button#manual-download-button.btn.btn-primary(hidden="false", type='button', style='width: 100%;') Continue
-  script.
-    ipcRenderer.on("user:manual_download", (event, file, group) => {
-      views.show("manual-download");
-      shell.openExternal(file.url);
-      $("#manual-download-group").text(group);
-      $("#manual-download-filename-1").text(file.name);
-      $("#manual-download-filename-2").text(file.name);
-      $("#manual-download-link").text(file.url);
-      $("#manual-download-link").one("click", () => {
-        shell.openExternal(file.url);
-      });
-
-      // HACK remove once https://github.com/electron/electron/pull/19159 is released
-      if (global.packageInfo.package === "snap") {
-        $("#manual-download-after-instructions").text("Once you have it, enter the path and click the button below to continue.");
-        document.getElementById("manual-download-input").type = "";
-        document.getElementById("manual-download-input").placeholder = "enter path here";
-      }
-
-      $("#manual-download-button").one("click", () => {
-        event.sender.send(
-          "manual_download:completed",
-          // HACK remove once https://github.com/electron/electron/pull/19159 is released
-          (global.packageInfo.package !== "snap") ?
-          ($("#manual-download-input")[0].files[0].path) :
-          ($("#manual-download-input")[0].value)
-        );
-      });
-    });
