--- src/pug/views/select-os.pug
@@ -1,38 +1,0 @@
-#views-select-os.main.container.views(hidden='hidden')
-  .row
-    .col-xs-6
-      img(style='height: 350px; margin: auto; display: block;', src='../screens/Screen6.jpg')
-    .col-xs-6(style='height: 100%')
-      h4.user-install-header#device-name(style='font-weight: bold;')
-      p #[a#device-page-link about this device] | #[a#device-config view config file]
-      p Please make sure you enabled #[a(onclick="modals.show('developer-mode-info')") developer mode and OEM unlocking].
-      p What operating system do you want to install?
-      form.form-horizontal
-        .form-group
-          .col-xs-3
-            label.control-label OS
-          .col-xs-9
-            select#options-os.form-control.space
-      button#btn-inst.btn.btn-primary(type='button', style='width: 100%; margin-top: 10px;') Install
-  script.
-    $("#btn-inst").click(() => {
-      views.show("working", "particles");
-      ipcRenderer.send(
-        "os:selected",
-        document.getElementById('options-os').selectedIndex
-      );
-    })
-
-    ipcRenderer.on("user:os", (event, config, osSelects) => {
-      footer.topText.set(config.name+" ("+config.codename+")");
-      footer.underText.set("Please select an operating system for installation");
-      $("#options-os").append(osSelects);
-      $("#device-name").text(config.name+" ("+config.codename+")");
-      $("#device-page-link").click(() => {
-        shell.openExternal("https://devices.ubuntu-touch.io/device/" + config.codename);
-      });
-      $("#device-config").click(() => {
-        shell.openExternal("https://github.com/ubports/installer-configs/blob/master/v1/" + config.codename + ".json");
-      });
-      views.show("select-os");
-    });
