--- src/pug/views/working.pug
@@ -1,10 +1,0 @@
-#views-working.vertical-centered-box.views(hidden='hidden')
-  .content
-    .loader-circle
-    .loader-line-mask
-      .loader-line
-img.ubp-robot(width='100', src='../img/yumi.svg')
-script.
-  ipcRenderer.on("user:write:working", (e, animation) => {
-    views.show("working", animation);
-  });
