+++ src/ui/views/UserAction.svelte
@@ -19,10 +19,10 @@
 </script>
 
 <div class="row">
-    <div class="col-xs-6">
+    <div class="col-6">
         <img src={userAction_Image?`./img/${userAction_Image}.jpg`:"./screens/Screen6.jpg"} alt="Screen6" style='height: 350px; margin: auto; display: block;'>
     </div>
-    <div class="col-xs-6">
+    <div class="col-6">
         <h4 style='font-weight: bold;'>
             {userAction_Title}
         </h4>
